Remove extra cast

+++ Dalamud/Game/NativeUi/OverlayAddon.cs
@@ -77,7 +77,7 @@ internal class OverlayAddon : NativeAddon
 
         foreach (var node in this.queuedNodes)
         {
-            node.PerformAttach((nint)this.InternalAddon);
+            node.PerformAttach(this.InternalAddon);
             this.attachedNodes.Add(node);
         }
